Fix issue causing make dependencies to degenerate on successive builds
Because all make dependencies (including headers) are globbed into the
gcc invocation, the dependencies of targets (-MMD output) will be
reduced on each invocation. Because the headers were specified on the
compiler invocation, they are no longer added with an #include
directive, and then gcc stops reporting them as dependencies.